requirements.txt
apache-airflow==2.9.3
dvc[gs]==3.50.0
datasets==2.19.0
langdetect==1.0.9
tqdm==4.66.0
pandas==2.2.0
google-cloud-storage==2.16.0
ftfy==6.2.0
sacrebleu==2.4.0
